fix(major::pr): re-architect instrumentation library
